Movement of robot 1
This is the code that causes the robot to move. It uses Pythagoras' theorem which means the path can change depending on where the treasure is and multiple paths do not need to be manually coded. This part of the code represents if x01>0 where x01 is the difference between the robot and the treasure in the x axis.